requirements.txt
flask
flask_sqlalchemy
pgvector
psycopg2-binary
requests
redis
pandas
sentence-transformers
pytest
openai